[FEAT] 모집게시글 ui 백연동
