+++ server/mockup.json
@@ -1277,197 +1277,3 @@
     }
   ]
 }
-    "profiles": [
-        {    
-            "id": 1,
-            "nickname": "관리자",
-            "email": "help@tododdudo.com",
-            "image": "/images/admin-profile.png"
-        },
-        {
-            "id": 2,
-            "nickname": "삥삥이",
-            "email": "bbing@mail.com",
-            "image": "/images/header-profile.png"
-        }
-    ],
-    "reports": [
-        {
-            "id": 9,
-            "type": "게시글",
-            "date": "2025-04-23",
-            "reporter": "삥삥이",
-            "content": "상업적/홍보성",
-            "status": "처리중",
-            "isProcessed": true,
-            "isApproved": false
-          },
-          {
-            "id": 8,
-            "type": "회원",
-            "date": "2025-04-22",
-            "reporter": "도고미",
-            "content": "욕설/인신공격",
-            "status": "처리중",
-            "isProcessed": true,
-            "isApproved": false
-          },
-          {
-            "id": 7,
-            "type": "게시글",
-            "date": "2025-04-21",
-            "reporter": "하비루",
-            "content": "음란/선정성",
-            "status": "처리중",
-            "isProcessed": true,
-            "isApproved": false
-          },
-          {
-            "id": 6,
-            "type": "게시글",
-            "date": "2025-04-20",
-            "reporter": "제리",
-            "content": "개인정보노출",
-            "status": "처리중",
-            "isProcessed": true,
-            "isApproved": false
-          },
-          {
-            "id": 5,
-            "type": "회원",
-            "date": "2025-04-19",
-            "reporter": "랑랑쓰",
-            "content": "부적절한 닉네임",
-            "status": "처리중",
-            "isProcessed": true,
-            "isApproved": false
-          },
-          {
-            "id": 4,
-            "type": "회원",
-            "date": "2025-04-18",
-            "reporter": "로지",
-            "content": "욕설/인신공격",
-            "status": "처리중",
-            "isProcessed": true,
-            "isApproved": false
-          },
-          {
-            "id": 3,
-            "type": "게시글",
-            "date": "2025-04-17",
-            "reporter": "토미",
-            "content": "불법정보",
-            "status": "처리중",
-            "isProcessed": true,
-            "isApproved": false
-          },
-          {
-            "id": 2,
-            "type": "회원",
-            "date": "2025-04-16",
-            "reporter": "도리",
-            "content": "음란/선정성",
-            "status": "처리중",
-            "isProcessed": true,
-            "isApproved": false
-          },
-          {
-            "id": 1,
-            "type": "게시글",
-            "date": "2025-04-15",
-            "reporter": "삥삥이",
-            "content": "기타",
-            "status": "처리중",
-            "isProcessed": true,
-            "isApproved": false
-          }
-    ],
-    "todoList" : [
-      { "id": 1, "text": "아침 7시에 일어나기" },
-      { "id": 2, "text": "하루에 30분 자격증 공부하기" }
-    ],
-    "approveList" : [
-      { "id": 1, "requesterName": "CANDDY" },
-      { "id": 2, "requesterName": "OKKZZI" }
-    ],
-    
-      "rooms": [
-        { "id": 1, "title": "공부습관방", "content": "공부습관을 만드는 사람들의 모임입니다." },
-        { "id": 2, "title": "아침형인간", "content": "매일 아침 일찍 일어나 도전하는 방입니다." },
-        { "id": 3, "title": "헬스메이트", "content": "운동 인증과 헬스 루틴을 함께 공유합니다." },
-        { "id": 4, "title": "다이어트파티", "content": "다이어트 식단과 운동을 서로 응원하는 방입니다." },
-        { "id": 5, "title": "하루만 공부해", "content": "하루만! 단 하루만 공부하기로 약속하는 방입니다." }
-      ]
-    
-    
-    ],
-    "post": [
-        {
-          "id": 1,
-          "category": "일상",
-          "title": "같이 매일 3시간씩 뉴스 요약하실 분 구합니다",
-          "status": "모집중"
-        },
-        {
-          "id": 2,
-          "category": "건강",
-          "title": "같이 매일 3시간씩 운동하실 분 구합니다",
-          "status": "모집중"
-        },
-        {
-          "id": 3,
-          "category": "자기개발",
-          "title": "같이 매일 3시간씩 코딩 공부하실 분 구합니다",
-          "status": "모집중"
-        },
-        {
-          "id": 4,
-          "category": "학습",
-          "title": "같이 매일 3시간씩 자격증 준비하실 분 구합니다",
-          "status": "모집중"
-        },
-        {
-          "id": 5,
-          "category": "학습",
-          "title": "같이 매일 3시간씩 일본어 공부하실 분 구합니다",
-          "status": "모집중"
-        },
-        {
-          "id": 6,
-          "category": "자기개발",
-          "title": "같이 매일 1시간씩 창작 글쓰기 하실 분 구합니다",
-          "status": "모집중"
-        },
-        {
-          "id": 7,
-          "category": "일상",
-          "title": "같이 매일 3시간씩 블로그 쓰실 분 구합니다",
-          "status": "모집중"
-        },
-        {
-          "id": 8,
-          "category": "건강",
-          "title": "같이 매일 3시간씩 플랭크 챌린지 하실 분 구합니다",
-          "status": "모집중"
-        },
-        {
-          "id": 9,
-          "category": "학습",
-          "title": "같이 매일 3시간씩 독서하실 분 구합니다",
-          "status": "모집마감"
-        },
-        {
-          "id": 10,
-          "category": "문화",
-          "title": "같이 매일 3시간씩 그림 그리실 짝꿍 구합니다",
-          "status": "모집중"
-        }
-      ]
-      
-
-
-
-
-
-}
